pyproject.toml
[project]
name = "xml-hwpx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-hwpx>=2.5",
]
